+++ test/slim_data_uri_test.dart
@@ -32,6 +32,33 @@ void main() {
       expect(slimUri.data.contentAsBytes(), equals(bytes));
     });
 
+    test('Parse - no safety check', () {
+      final malicious = 'DUMMY"/><script>alert("attack!");</script>';
+      final slimUri = SlimDataUri.parse('data:;base64,$malicious');
+      expect(slimUri.toString(), equals('data:;base64,$malicious'));
+      expect(slimUri.path, equals(';base64,$malicious'));
+      expect(slimUri.scheme, equals('data'));
+      expect(slimUri.isSafeContent, isFalse);
+      expect(slimUri.data.isBase64, isTrue);
+      expect(slimUri.data.mimeType, equals('application/octet-stream'));
+      expect(slimUri.data.contentAsBytes, throwsA(isA<Exception>()));
+    });
+
+    test('Parse - with safety check', () {
+      final malicious = 'DUMMY"/><script>alert("attack!");</script>';
+      expect(
+          () => SlimDataUri.parse('data:;base64,$malicious', safetyCheck: true),
+          throwsA(isA<Exception>()));
+    });
+
+    test('Parse - safe yet invalid payload', () {
+      final invalid = 'A=B';
+      final slimUri =
+          SlimDataUri.parse('data:;base64,$invalid', safetyCheck: true);
+      expect(slimUri.isSafeContent, isTrue);
+      expect(slimUri.data.contentAsBytes, throwsA(isA<Exception>()));
+    });
+
     test('Dart Uri --> SlimDataUri', () {
       final uri = Uri.parse('data:application/x-test;base64,$base64');
       final slimUri = SlimDataUri.parse(uri.toString());
@@ -135,6 +162,31 @@ void main() {
       expect(slimUri.data.contentAsBytes(), equals(bytes));
     });
 
+    test('Parse - no safety check', () {
+      final malicious = 'DUMMY"/><script>alert("attack!");</script>';
+      final slimUri = SlimDataUri.parse('data:,$malicious');
+      expect(slimUri.toString(), equals('data:,$malicious'));
+      expect(slimUri.path, equals(',$malicious'));
+      expect(slimUri.scheme, equals('data'));
+      expect(slimUri.isSafeContent, isFalse);
+      expect(slimUri.data.isBase64, isFalse);
+      expect(slimUri.data.mimeType, equals('application/octet-stream'));
+      expect(slimUri.data.contentAsBytes, throwsA(isA<Exception>()));
+    });
+
+    test('Parse - with safety check', () {
+      final malicious = 'DUMMY"/><script>alert("attack!");</script>';
+      expect(() => SlimDataUri.parse('data:,$malicious', safetyCheck: true),
+          throwsA(isA<Exception>()));
+    });
+
+    test('Parse - safe yet invalid payload', () {
+      final invalid = '%%';
+      final slimUri = SlimDataUri.parse('data:,$invalid', safetyCheck: true);
+      expect(slimUri.isSafeContent, isTrue);
+      expect(slimUri.data.contentAsBytes, throwsA(isA<Exception>()));
+    });
+
     test('Dart Uri --> SlimDataUri', () {
       final uri = Uri.parse('data:application/x-test,$percent');
       final slimUri = SlimDataUri.parse(uri.toString());
@@ -225,7 +277,6 @@ final bytes = [
   0x41,
   0x61,
   0x0A,
-  0x7E,
   0x7F,
   0x80,
   0x81,
@@ -234,6 +285,7 @@ final bytes = [
   0xFF,
   0x20, // space
   0x21, // !
+  0x22, // "
   0x23, // #
   0x24, // $
   0x25, // %