+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+
+- Add optional `safetyCheck` parameter (default value: `false`) to check for payload safety upon construction.
+- Add property `isSafeContent` property to check for payload safety.
+
 ## 1.1.1
 
 - Update README with security considerations.